docker/import_check.py
import cv2
import nvdiffrast.torch as dr
import open3d as o3d
import torch
import trimesh
import warp as wp
import mycpp.build.mycpp as mycpp
from bundlesdf.mycuda import common, gridencoder

print("torch:", torch.__version__)
print("cuda_available:", torch.cuda.is_available())
print("foundationpose_import_check: ok")
