docker/runtime-requirements.txt
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.4.1.post1
scikit-image==0.22.0
pyyaml==6.0.1
ruamel.yaml==0.18.6
ninja==1.11.1.1
h5py==3.10.0
numba==0.59.1
pybind11==2.12.0
imageio==2.34.0
opencv-contrib-python==4.9.0.80
open3d==0.18.0
trimesh==4.2.2
pyrender==0.1.45
pyOpenGL>=3.1.0
pyOpenGL_accelerate>=3.1.0
omegaconf==2.3.0
pyglet==1.5.28
kornia==0.7.2
einops==0.7.0
transformations==2024.6.1
joblib==1.3.2
warp-lang==1.0.2
fvcore==0.1.5.post20221221
pandas==2.0.3
matplotlib==3.7.5
psutil==5.9.8
Pillow==10.2.0
tqdm==4.66.2
